# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI Searcher API",
    "description": "Запрос к нескольким LLM-провайдерам за один вызов: **`POST /api/query`** (JSON целиком) или **`POST /api/query/stream`** (SSE, по одной ячейке). **`GET /api/meta`** — какие модели настроены в `.env` (без секретов). Если задан `API_SECRET` — **Authorization: Bearer** или **X-API-Key** только для POST.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://gpt.seo-performance.ru",
      "description": "Продакшен"
    },
    {
      "url": "http://localhost:3847",
      "description": "Локально (порт по умолчанию из `PORT`)"
    }
  ],
  "tags": [
    {
      "name": "Query",
      "description": "Поиск по моделям"
    }
  ],
  "paths": {
    "/api/meta": {
      "get": {
        "tags": ["Query"],
        "summary": "Список моделей и настройка в .env",
        "description": "Публично, **без** заголовка `API_SECRET`. Для UI: какие провайдеры доступны по ключам на сервере; поле **`proxy`** — задан ли для провайдера исходящий HTTP(S)-прокси (без URL).",
        "operationId": "getMeta",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MetaResponse" }
              }
            }
          }
        }
      }
    },
    "/api/query": {
      "post": {
        "tags": ["Query"],
        "summary": "Запрос к выбранным моделям",
        "description": "Либо одно поле **`query`**, либо массив **`queries`** (пакет до 120 строк). Пустые строки в `queries` отбрасываются. Поле **`providers`**: `[\"all\"]` или список id. Провайдеры **без ключа в .env** всё равно попадают в **`results`** с **`error`** и пустым **`text`**. **`skippedLabels`** — только подписи отключённых моделей из **текущего** списка `providers`.",
        "operationId": "postQuery",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  { "$ref": "#/components/schemas/SingleQueryRequest" },
                  { "$ref": "#/components/schemas/BatchQueryRequest" }
                ]
              },
              "examples": {
                "singleAll": {
                  "summary": "Один запрос, все настроенные модели",
                  "value": {
                    "query": "лучшее SEO-агентство в Москве",
                    "providers": ["all"]
                  }
                },
                "singleSubset": {
                  "summary": "Один запрос, две модели",
                  "value": {
                    "query": "что такое structured data",
                    "providers": ["chatgpt", "google"]
                  }
                },
                "batch": {
                  "summary": "Пакет запросов",
                  "value": {
                    "queries": ["первый запрос", "второй запрос"],
                    "providers": ["all"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успех (одиночный или пакетный ответ)",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "$ref": "#/components/schemas/SingleQueryResponse" },
                    { "$ref": "#/components/schemas/BatchQueryResponse" }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Пустой/длинный запрос, нет провайдеров, ошибка валидации",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "empty": { "value": { "error": "Пустой запрос" } },
                  "noProviders": {
                    "value": {
                      "error": "Ни один провайдер не выбран или не настроен (проверьте .env).",
                      "results": [],
                      "skippedLabels": ["ChatGPT", "DeepSeek"]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "На сервере задан `API_SECRET`, заголовок авторизации неверный или отсутствует",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/query/stream": {
      "post": {
        "tags": ["Query"],
        "summary": "Поток результатов (Server-Sent Events)",
        "description": "Тело запроса **то же**, что у `POST /api/query`: поле **`query`** или массив **`queries`** (до 120 строк) и **`providers`**. При успехе ответ — **`text/event-stream`**: события SSE в виде строк `data: ` + JSON (см. **`StreamEvent`**) + `\\n\\n`. Обычный порядок: **`meta`** → несколько **`result`** → **`done`**. Если ни один провайдер не настроен, может уйти только **`error`**, затем соединение закрывается. Заголовки: `Cache-Control: no-cache, no-transform`, `Connection: keep-alive`, `X-Accel-Buffering: no`. Ошибки **400** и **401** до открытия потока — обычный **`application/json`**, не SSE.",
        "operationId": "postQueryStream",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  { "$ref": "#/components/schemas/SingleQueryRequest" },
                  { "$ref": "#/components/schemas/BatchQueryRequest" }
                ]
              },
              "examples": {
                "batchStream": {
                  "summary": "Пакет (как в UI)",
                  "value": {
                    "queries": ["запрос один", "запрос два"],
                    "providers": ["all"]
                  }
                },
                "singleStream": {
                  "summary": "Один запрос",
                  "value": {
                    "query": "кратко: что такое structured data",
                    "providers": ["chatgpt", "google"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Поток SSE. Тело — последовательность блоков `data: <JSON>\\n\\n`; тип события в поле **`type`** (схема **`StreamEvent`**).",
            "headers": {
              "Cache-Control": {
                "description": "Отключение кэширования для живого потока",
                "schema": { "type": "string", "example": "no-cache, no-transform" }
              },
              "X-Accel-Buffering": {
                "description": "Для nginx: не буферизовать SSE",
                "schema": { "type": "string", "example": "no" }
              }
            },
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "Формат SSE: повторяющиеся фрагменты `data: ` + один объект JSON (`StreamEvent`) + перевод строки; блок завершается пустой строкой."
                },
                "example": "data: {\"type\":\"meta\",\"queries\":[\"привет\"],\"providerIds\":[\"chatgpt\"],\"totalTasks\":1,\"skippedLabels\":[\"Perplexity\"]}\n\ndata: {\"type\":\"result\",\"queryIndex\":0,\"providerId\":\"chatgpt\",\"progress\":100,\"result\":{\"id\":\"chatgpt\",\"label\":\"ChatGPT\",\"text\":\"Ответ…\",\"links\":[\"https://example.com\"],\"error\":null,\"durationMs\":1200,\"usage\":{\"input\":10,\"output\":20,\"total\":30}}}\n\ndata: {\"type\":\"done\",\"skippedLabels\":[\"Perplexity\"]}\n\n"
              }
            }
          },
          "400": {
            "description": "Ошибка валидации до открытия потока (JSON)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Неверный или отсутствующий секрет при заданном `API_SECRET` (JSON)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка (редко как JSON до потока; при сбое во время потока сервер может записать SSE с `type: error`)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Тот же секрет, что в `API_SECRET` (если переменная задана). Иначе заголовок не нужен."
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Альтернатива Bearer; значение = `API_SECRET`."
      }
    },
    "schemas": {
      "MetaProviderEntry": {
        "type": "object",
        "required": ["id", "label", "configured", "proxy"],
        "properties": {
          "id": { "type": "string", "description": "Id провайдера" },
          "label": { "type": "string", "description": "Подпись для UI" },
          "configured": {
            "type": "boolean",
            "description": "true, если на сервере заданы нужные переменные .env для этого провайдера"
          },
          "proxy": {
            "type": "boolean",
            "description": "true, если для этого провайдера задан URL исходящего HTTP(S)-прокси в .env (без раскрытия URL)"
          }
        }
      },
      "MetaResponse": {
        "type": "object",
        "required": ["providers"],
        "properties": {
          "providers": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/MetaProviderEntry" }
          }
        }
      },
      "ProviderId": {
        "type": "string",
        "enum": ["all", "chatgpt", "deepseek", "perplexity", "google", "alice", "alice_search"]
      },
      "ProvidersSelection": {
        "type": "array",
        "items": { "$ref": "#/components/schemas/ProviderId" },
        "description": "Пустой или неизвестный id → как `['all']`."
      },
      "TokenUsage": {
        "type": "object",
        "nullable": true,
        "properties": {
          "input": { "type": "integer", "nullable": true, "description": "Prompt / вход" },
          "output": { "type": "integer", "nullable": true, "description": "Completion / выход" },
          "total": { "type": "integer", "nullable": true }
        }
      },
      "ProviderResult": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "label": { "type": "string" },
          "text": { "type": "string" },
          "links": {
            "type": "array",
            "items": { "type": "string", "format": "uri" }
          },
          "usage": { "$ref": "#/components/schemas/TokenUsage" },
          "error": {
            "type": "string",
            "description": "Ошибка вызова модели или провайдер не настроен на сервере (нет ключа в .env) — тогда в `results` всё равно есть объект с `id`/`label` и пустым `text`."
          },
          "durationMs": { "type": "integer" }
        }
      },
      "SingleQueryRequest": {
        "type": "object",
        "required": ["query"],
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "maxLength": 8000,
            "description": "Текст запроса"
          },
          "providers": { "$ref": "#/components/schemas/ProvidersSelection" }
        }
      },
      "BatchQueryRequest": {
        "type": "object",
        "required": ["queries"],
        "properties": {
          "queries": {
            "type": "array",
            "items": { "type": "string", "minLength": 1, "maxLength": 8000 },
            "minItems": 1,
            "maxItems": 120,
            "description": "Каждый элемент — отдельный пользовательский запрос"
          },
          "providers": { "$ref": "#/components/schemas/ProvidersSelection" }
        }
      },
      "SingleQueryResponse": {
        "type": "object",
        "required": ["results", "skippedLabels"],
        "properties": {
          "results": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ProviderResult" }
          },
          "skippedLabels": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Подписи провайдеров без ключа в .env"
          },
          "error": { "type": "string", "nullable": true }
        }
      },
      "BatchItem": {
        "type": "object",
        "properties": {
          "query": { "type": "string" },
          "error": { "type": "string", "nullable": true },
          "results": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ProviderResult" }
          }
        }
      },
      "BatchQueryResponse": {
        "type": "object",
        "required": ["batch", "items", "skippedLabels"],
        "properties": {
          "batch": { "type": "boolean", "enum": [true] },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/BatchItem" }
          },
          "skippedLabels": {
            "type": "array",
            "items": { "type": "string" }
          }
        }
      },
      "StreamEventMeta": {
        "type": "object",
        "required": ["type", "queries", "providerIds", "totalTasks", "skippedLabels"],
        "properties": {
          "type": { "type": "string", "enum": ["meta"] },
          "queries": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Список пользовательских запросов в порядке обработки"
          },
          "providerIds": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Id активных провайдеров для этого прогона"
          },
          "totalTasks": {
            "type": "integer",
            "minimum": 0,
            "description": "Число ячеек: запросы × модели (ожидаемое число событий `result`)"
          },
          "skippedLabels": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Подписи провайдеров без ключа в .env"
          }
        }
      },
      "StreamEventResult": {
        "type": "object",
        "required": ["type", "queryIndex", "providerId", "progress", "result"],
        "properties": {
          "type": { "type": "string", "enum": ["result"] },
          "queryIndex": {
            "type": "integer",
            "minimum": 0,
            "description": "Индекс строки в `meta.queries`"
          },
          "providerId": { "type": "string", "description": "Id модели" },
          "progress": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "description": "Процент завершённых ячеек (0–100)"
          },
          "result": {
            "$ref": "#/components/schemas/ProviderResult",
            "description": "Ответ одной модели для одной ячейки"
          }
        }
      },
      "StreamEventDone": {
        "type": "object",
        "required": ["type", "skippedLabels"],
        "properties": {
          "type": { "type": "string", "enum": ["done"] },
          "skippedLabels": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Те же пропущенные модели, что в `meta`"
          }
        }
      },
      "StreamEventError": {
        "type": "object",
        "required": ["type", "message"],
        "properties": {
          "type": { "type": "string", "enum": ["error"] },
          "message": { "type": "string", "description": "Текст ошибки (например, нет провайдеров)" },
          "skippedLabels": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Опционально, при ошибке выбора провайдеров"
          }
        }
      },
      "StreamEvent": {
        "description": "Один JSON в строке `data:` SSE",
        "oneOf": [
          { "$ref": "#/components/schemas/StreamEventMeta" },
          { "$ref": "#/components/schemas/StreamEventResult" },
          { "$ref": "#/components/schemas/StreamEventDone" },
          { "$ref": "#/components/schemas/StreamEventError" }
        ],
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "meta": "#/components/schemas/StreamEventMeta",
            "result": "#/components/schemas/StreamEventResult",
            "done": "#/components/schemas/StreamEventDone",
            "error": "#/components/schemas/StreamEventError"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "results": { "type": "array", "nullable": true },
          "skippedLabels": { "type": "array", "items": { "type": "string" }, "nullable": true },
          "batch": { "type": "boolean", "nullable": true },
          "items": { "type": "array", "nullable": true }
        }
      }
    }
  }
}